+++ doc-examples/src/main/java/org/realityforge/arez/doc/examples/multi_zone/Example.java
@@ -6,16 +6,16 @@ import org.realityforge.arez.Zone;
 public class Example
 {
   public static void main( final String[] args )
+    throws Throwable
   {
     // Create zone
     final Zone zone = Arez.createZone();
     // Activate the newly created zone and suspend the current zone if any
-    zone.activate();
+    zone.run( () -> {
+      // Create new component and ensure it is bound to current zone
+      final MyComponent myComponent = new Arez_MyComponent();
 
-    // Create new component and ensure it is bound to current zone
-    final MyComponent myComponent = new Arez_MyComponent();
-
-    // Deactivate the zone and restore the previous if any
-    zone.deactivate();
+      // Zone is deactivated when leaves this block
+    } );
   }
 }